requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
httpx>=0.27.0
google-genai>=1.60.0
google-auth>=2.29.0
numpy>=1.26.4
asyncpg>=0.29.0
SQLAlchemy[asyncio]>=2.0.0
alembic>=1.13.0
PyJWT>=2.8.0
slowapi>=0.1.9
google-api-python-client>=2.149.0
sse-starlette>=2.0.0